backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg==3.1.18
gunicorn==21.2.0
Pillow==10.2.0
whitenoise==6.6.0